Moved files. Added new icons. Added text by build button. Added gold and player resources. Hooked UI to player resources.

+++ Assets/Scripts/Buildings/BuildingData.cs
@@ -1,5 +1,4 @@
 ﻿using UnityEngine;
-using UnityEngine.UI;
 
 [CreateAssetMenu]
 public class BuildingData : ScriptableObject
@@ -7,6 +6,6 @@ public class BuildingData : ScriptableObject
 	public BuildingType Type;
 	public string Name;
 	public Sprite Icon;
-	public float Cost;
-	public float ConstructionTime;
+	public int Cost;
+	public int ConstructionTime;
 }
